refactor the get model function
